Extract MediaCache from MediaCacheStorage.kt, Move TestMediaCacheStorage
